requirements.txt
# Core dependencies
boto3>=1.34.0
python-dotenv>=1.0.0
qdrant-client>=1.7.0

# Dataset loading (for wiki.py)
datasets>=2.14.0

# Tigris-specific extensions
tigris-boto3-ext>=0.1.0

# CrewAI (optional, for crew.py)
crewai>=0.28.0
crewai-tools>=0.1.0

# Streamlit (for streamlit_app.py)
streamlit>=1.28.0

